✨ page 폴더 내부 로직 분리
