+++ src/entities/tag/api/api.ts
@@ -1,0 +1,3 @@
+import axios from 'axios';
+
+export const getPostTags = async () => await axios.get('/api/posts/tags');
